# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AirGo前后分离,多用户代理面板",
        "title": "AirGo",
        "contact": {
            "url": "https://github.com/pura-panel/airgo-panel"
        },
        "license": {
            "name": "GPL v3.0",
            "url": "https://github.com/pura-panel/airgo-panel/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/api/admin/order/getOrderList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] order"
                ],
                "summary": "获取全部订单，分页获取",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/order/orderSummary": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] order"
                ],
                "summary": "获取订单统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/order/updateOrder": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] order"
                ],
                "summary": "更新用户订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
...
        "/api/customer/finance/getBalanceStatementList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] finance"
                ],
                "summary": "获取余额流水",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/finance/getCommissionStatementList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] finance"
                ],
                "summary": "获取佣金流水",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/finance/getCommissionSummary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] finance"
                ],
                "summary": "获取佣金统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/finance/getInvitationUserList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] finance"
                ],
                "summary": "获取邀请人数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/finance/withdrawToBalance": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] finance"
                ],
                "summary": "提现",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
...
        "/api/public/shop/epayNotify": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[public api] shop"
                ],
                "summary": "易支付异步回调",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EpayResultResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AlipayPreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "qr_code": {
                    "type": "string"
                }
            }
        },
...
        "model.EpayPreCreatePay": {
            "type": "object",
            "properties": {
                "money": {
                    "description": "商品金额,如：1.00\t单位：元，最大2位小数",
                    "type": "string"
                },
                "name": {
                    "description": "商品名称,如超过127个字节会自动截取",
                    "type": "string"
                },
                "notify_url": {
                    "description": "服务器异步通知地址",
                    "type": "string"
                },
                "out_trade_no": {
                    "description": "商户订单号",
                    "type": "string"
                },
                "pid": {
                    "description": "商户ID",
                    "type": "integer"
                },
                "return_url": {
                    "description": "页面跳转通知地址",
                    "type": "string"
                },
                "sign": {
                    "description": "签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign = md5 ( a=b\u0026c=d\u0026e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。",
                    "type": "string"
                },
                "sign_type": {
                    "description": "签名类型",
                    "type": "string"
                },
                "type": {
                    "description": "支付方式",
                    "type": "string"
                }
            }
        },
        "model.EpayPreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "epay_api_url": {
                    "type": "string"
                },
                "epay_pre_create_pay": {
                    "$ref": "#/definitions/model.EpayPreCreatePay"
                }
            }
        },
        "model.EpayResultResponse": {
            "type": "object",
            "properties": {
                "money": {
                    "description": "商品金额",
                    "type": "string"
                },
                "name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "out_trade_no": {
                    "description": "商户订单号",
                    "type": "string"
                },
                "param": {
                    "description": "业务扩展参数",
                    "type": "string"
                },
                "pid": {
                    "description": "商户ID",
                    "type": "integer"
                },
                "sign": {
                    "description": "签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign = md5 ( a=b\u0026c=d\u0026e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。",
                    "type": "string"
                },
                "sign_type": {
                    "description": "签名类型",
                    "type": "string"
                },
                "trade_no": {
                    "description": "易支付订单号",
                    "type": "string"
                },
                "trade_status": {
                    "description": "支付状态",
                    "type": "string"
                },
                "type": {
                    "description": "支付方式",
                    "type": "string"
                }
            }
        },
        "model.FieldParamsItem": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "= \u003e \u003c \u003c\u003e like",
                    "type": "string"
                },
                "condition_value": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "operator": {
                    "description": "AND OR",
                    "type": "string"
                }
            }
        },
...
        "model.Order": {
            "type": "object",
            "properties": {
                "balance_amount": {
                    "type": "string"
                },
                "buyer_logon_id": {
                    "type": "string"
                },
                "buyer_pay_amount": {
                    "type": "string"
                },
                "coupon_amount": {
                    "type": "string"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_service_id": {
                    "description": "服务参数",
                    "type": "integer"
                },
                "deliver_text": {
                    "type": "string"
                },
                "deliver_type": {
                    "type": "string"
                },
                "des": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "goods_id": {
                    "description": "商品参数",
                    "type": "integer"
                },
                "goods_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_remarks": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "original_amount": {
                    "type": "string"
                },
                "out_trade_no": {
                    "type": "string"
                },
                "pay_id": {
                    "description": "支付参数",
                    "type": "integer"
                },
                "pay_info": {
                    "description": "支付信息，epay，alipay等\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PreCreatePayToFrontend"
                        }
                    ]
                },
                "pay_type": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "string"
                },
                "trade_no": {
                    "type": "string"
                },
                "trade_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "关联用户",
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "model.Pagination": {
            "type": "object",
            "properties": {
                "order_by": {
                    "type": "string"
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "model.PreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "alipay_info": {
                    "$ref": "#/definitions/model.AlipayPreCreatePayToFrontend"
                },
                "epay_info": {
                    "$ref": "#/definitions/model.EpayPreCreatePayToFrontend"
                }
            }
        },
...
        "model.QueryParams": {
            "type": "object",
            "required": [
                "table_name"
            ],
            "properties": {
                "field_params_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldParamsItem"
                    }
                },
                "pagination": {
                    "description": "分页参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}